making the separated Data file
